(0.1.5) MODULARITE ESPACE UTILISATEUR
0.1.32 -> 0.1.5 => 0.1.6
[MODULAR | FILE . ++ | ORG . ++ | FUNC . ++]

Description/

Espace utilisateur fonctionnel

Header ajouté commun a toute les pages

connexion inscription fait , page de config avec gestion par liste fait

Nouveau moteur de Wilber prêt a acceuillir sereinement les futurs mise a jour

====================================================================
*/FILE

	* Création de la page d'inscription utilisateur
	* Création de la page espace utilisateur
	* dossier INC ajouté

====================================================================
*/DOC

====================================================================
*/MODULAR

	-> Créer le moyen de pouvoir poster différents commentaire depuis des pages différentes et les afficher 	selon une option choisit

	-> idem avec article

====================================================================
*/FUNC
	* Espace utilisateur fonctionnel avec lien de confirmation

	* Possibilité de suprrimer gérer les utilisateurs et les articles
	-> Possibilité de modification groupé en cours
	* Possibilité de connexion par le header

====================================================================
*/ORG

====================================================================
*/DESIGN

+++ Sql/WB_sql_treatment.php
@@ -4,6 +4,30 @@
 require "../Wilber.php";
 
 
+// POST STATMENT 
+
+if($_POST){
+    $posted_elements = toolbox::parse_post();
+    var_dump($posted_elements);
+
+    if(!empty($posted_elements)){
+        $data_statment = new formManager;
+        foreach($posted_elements as $posted_elements){
+            $data_statment->processing_data_form("article",$posted_elements);
+        }
+        
+    }else{
+        $data_statment = new formManager;
+        $data_statment->processing_data_form("article",$_POST);
+    }
+
+    die();
+}
+
+
+
+// GET STATMENT 
+
     if(isset($_GET['article_del_id'])){
         echo "oui";
         $q = DB::$pdo->prepare('DELETE FROM article WHERE id = ?');
@@ -33,4 +57,21 @@ require "../Wilber.php";
         $q->execute([$_GET['user_del_id']]);
 
     }
-    header("location: " . $_GET['url_origin']);
+
+    if(isset($_GET['confirmation_token'])){
+        $q = DB::$pdo->prepare("SELECT confirmation_token FROM user WHERE id = ? ");
+        $q->execute([$_GET['user_id']]);
+        $result = $q->fetch();
+
+
+        if($result->confirmation_token == $_GET['confirmation_token']){
+            $q = DB::$pdo->prepare('UPDATE USER SET confirmation_token = NULL,confirmed_at = NOW() WHERE id = ?')->execute([$_GET['user_id']]);
+            toolbox::write_flash_message("Felicitation ". userManager::connected_name()." Vous avez confirmé votre compte avec succès","success");
+        }
+    }
+    if(isset($_GET['disconect'])){
+        unset($_SESSION['logged_in']);
+        toolbox::write_flash_message("Vous avez été déconnecté avec succès ! ","success");
+    }
+    
+    header("location: ../WBP_accueil.php");
